Split the simulator UI into windowed UI
This takes the existing simulator UI and splits into two separate windows inside of a larger panel/windowed UI. The windowing system is handled by a library called react-mosiac-component, so I named things using the Mosiac to make it distinct from existing UIs. The kotlin still does all the rendering and updating of the HTML templates.

I added a simple global state to the MosiacUI that is redux-ish using react context in store.jsx. In order to get the SheepSimulator to update the 3D canvas, I had to make the visualizer field public. This is a little messy, but as we move the UI out of Kotlin control, things should get cleaner going forward.