Add a read-only GLSL show editor window
This is the start of what will become the live text based show editor. It uses react-ace, a react wrapper around the ace text editor library.

The editor uses the pubSub to listen for `selectedShow` changing and then snoops the AllShows class to read out the text for the show. For a full editor to work we’ll need some sort of persistance for the shows so we can edit and save them live.